Re-organize IngestView panels.
